Separated classes into separate files
